fix(vue-template)!: prioritize local variable over auto import
